pyproject.toml
[tool.poetry]
name = "database-searcher"
version = "0.1.0"
description = "This is a reusable UI that accesses and creates speadsheets of unique papers found from a searchterm in Pubmed and Embase."
authors = ["charbelmarche33 <cmarche@mail.umw.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
